AcornUtils/src/com/acornui/collection/CyclicList.kt
package com.acornui.collection


/**
 * A ring buffer with fast push/pop/unshift/shift.

 * Good for queues and stacks.
 * Does not support arbitrary insertion or removal, only push/pop/unshift/shift.

 * @author nbilyk
 */
class CyclicList<E>(initialCapacity: Int = 16) : Clearable, ListBase<E>() {

	private var items = ArrayList<E?>(initialCapacity)

	private var _size: Int = 0
	private var capacity: Int = initialCapacity

	override val size: Int
		get() = _size

	private var start: Int = 0

	init {
		items.fill(initialCapacity, { null })
	}

	/**
	 * pushFront
	 * Add each of the elements to beginning of the list. The added items will be in reverse order.
	 */
	fun unshift(vararg values: E) {
		for (i in 0..values.lastIndex) {
			unshift(values[i])
		}
	}

	/**
	 * pushFront
	 * Adds an element to the beginning of the list.
	 */
	fun unshift(value: E) {
		if (_size == capacity) {
			resize(maxOf(8, (_size * 1.75f).toInt()))
		}
		val localIndex = getLocalIndex(-1)
		items[localIndex] = value
		start = localIndex
		_size++
	}

	/**
	 * popFront
	 * Removes an item from the beginning of the list.
	 */
	fun shift(): E {
		if (_size == 0) throw Exception("List is empty")
		val item = items[start]!!
		items[start] = null
		start = getLocalIndex(1)
		_size--
		return item
	}

	/**
	 * pushBack
	 */
	fun addAll(values: Iterable<E>) {
		for (t in values) {
			add(t)
		}
	}

	/**
	 * pushBack
	 */
	fun addAll(vararg values: E) {
		for (i in 0..values.lastIndex) {
			add(values[i])
		}
	}

	/**
	 * pushBack
	 */
	fun add(value: E) {
		if (_size == capacity) {
			resize(maxOf(8, (_size * 1.75f).toInt()))
		}
		val localIndex = getLocalIndex(_size)
		items[localIndex] = value
		_size++
	}

	/**
	 * popBack
	 */
	fun pop(): E {
		if (_size == 0) throw Exception("List is empty")
		val localIndex = getLocalIndex(_size - 1)
		val item = items[localIndex]!!
		items[localIndex] = null
		_size--
		return item
	}

	override fun get(index: Int): E {
		val localIndex = getLocalIndex(index)
		return items[localIndex]!!
	}

	private fun getLocalIndex(index: Int): Int {
		var localIndex = start + index
		if (localIndex >= capacity) {
			localIndex -= capacity
			if (localIndex >= start) {
				throw IllegalArgumentException("Index is out of bounds: " + index)
			}
		}
		if (localIndex < 0) {
			localIndex += capacity
			if (localIndex < start + _size) {
				throw IllegalArgumentException("Index is out of bounds: " + index)
			}
		}
		return localIndex
	}

	override fun clear() {
		for (i in 0.._size - 1) {
			var localIndex = i + start
			if (localIndex >= capacity)
				localIndex -= capacity
			else if (localIndex < 0)
				localIndex += capacity
			items[localIndex] = null
		}
		_size = 0
		start = 0
	}

	/**
	 * Resizes the backing items array to meet the new capacity.
	 */
	private fun resize(newCapacity: Int) {
		val newItems = ArrayList<E?>(newCapacity)
		newItems.fill(newCapacity, { null })
		if (start + size > capacity) {
			// Items wrap.
			arrayCopy(items, start, newItems, 0, capacity - start)
			arrayCopy(items, 0, newItems, capacity - start, size - capacity + start)
		} else {
			arrayCopy(items, start, newItems, 0, size)
		}

		this.items = newItems
		this.start = 0
		this.capacity = newCapacity
	}
}

val cyclicListPool = ClearableObjectPool<CyclicList<*>> { CyclicList<Any>() }
